warning
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@20431 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMAll/REMAll.cpp
@@ -100,7 +100,7 @@ static void remNotifyHandlerInsert(PVM pVM, PREMHANDLERNOTIFICATION pRec)
     do
     {
         idxFree = pVM->rem.s.idxFreeList;
-        if (idxFree == -1)
+        if (idxFree == (uint32_t)-1)
         {
             pFree = NULL;
             break;
